REFACTOR: Move precompiled interface out of main
Main now has the "from source" version in it.
This is because it has the least dependencies.

To use the precompiled version, run
using Pkg
Pkg.add(url=REPO_URL, rev="precompiled")

There are three main ways of using this package:

1. Without Madagascar at all
    a. Pkg.add("Madagascar")
    b. This supports only `rsf_write` and `rsf_read`

2. With a local Madagascar installation
    a. Pkg.add("Madagascar")
    b. Assumes that Madagascar binaries are in RSFROOT env var

3. With precompiled binaries
    a. Pkg.add(url=REPO_URL, rev="precompiled")
    b. This is the most convenient way to use the package, but the least flexible.

The decision behind this refactor is because
    maintenance of the precompiled binaries is harder
    and the pre-existing build system of Madagascar is
    already robust.

Furthermore, this allows the package to be used
    in a more flexible way by allowing devs to
    use their own Madagascar binaries if they want.